Added getting of the current working attributes record immediately after the retrieve.
This resolves an issue of the attributes record being retrieve once as
the BC is registered and then never again. The attributes record
actually used at the time of the retrieve may have changed, so
re-getting it lets other programs then do a
BCBinder::getAttributesRecord and know it corresponds to the result set.