Catching up to the 'master' branch.
Adds getFilter() to IConfigurationsBC.

Sets EDITABLE attribute in createAttributesRecord() in SqlTableBC,
depending on whether the field is in the metadata or contained in the
primary key. Also adds toString() method.

Merges or adds a data row in the write() method in BCBinder, instead of
removing it, if only one record is selected.